created a createoffice endpoint
