+++ app/api/version1/views/party_views.py
@@ -26,22 +26,3 @@ def get_all_parties():
     return make_response(jsonify({
         'message': res
     }), 200)
-
-
-@v1.route('/createoffice', methods=['POST'])
-def create_offices():
-    office = request.get_json()
-    office_type = ['type']
-    office_name = ['name']
-    
-
-    new_office = {
-        'type': office_type,
-        'name': office_name
-        
-    }
-    parties.append(party)
-    Party.create_party(new_party)
-    return make_response(jsonify({
-        'message': 'Office added successfully'
-    }), 200)
